+++ MODULE_GLOB.test.mjs
@@ -18,8 +18,16 @@ export default (tests) => {
           "./test/fixtures/MODULE_GLOB/matching-modules/",
           import.meta.url
         ),
+        dot: true,
       }),
       [
+        ".a.ad.cts",
+        ".a.ad.mts",
+        ".a.cjs",
+        ".a.cts",
+        ".a.js",
+        ".a.mjs",
+        ".a.mts",
         ".cjs",
         ".cts",
         ".js",
@@ -28,6 +36,13 @@ export default (tests) => {
         "a.mjs",
         "ad.cts",
         "ad.mts",
+        "a/.a.ad.cts",
+        "a/.a.ad.mts",
+        "a/.a.cjs",
+        "a/.a.cts",
+        "a/.a.js",
+        "a/.a.mjs",
+        "a/.a.mts",
         "a/.cjs",
         "a/.cts",
         "a/.js",
@@ -47,6 +62,7 @@ export default (tests) => {
           "./test/fixtures/MODULE_GLOB/excluded-files/",
           import.meta.url
         ),
+        dot: true,
       }),
       []
     );